Refactor load-file-content.js into focused modules
The 250-line file was doing five unrelated things. Split into:
- editing/manage-unsaved-changes.js — tracks whether content has changed since open
- editing/history-navigation.js — switches modal between current and historical versions
- ui-functions-render/render-file-content.js — renders modal content and unsaved indicator
- ui-functions-click/toggle-render-text.js — handles the txt/html mode toggle action
- ui-functions-click/file-content-input.js — handles the contentEditable input event

Content session state (activeRaw, liveRaw, isDirty, etc.) moved to appState.editSession
in store.js so the new modules can share it without coupling to each other.
